requirements.txt
flask==3.1.2
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.43
gunicorn==23.0.0
authlib==1.6.4
alembic==1.16.5
python-dotenv==1.1.1
flask-wtf==1.2.1
flask-limiter==3.8.0
bleach==6.1.0
flask-mail==0.9.1
